Add header parsing function, introduce monitoring routes, and define new types for job data and monitoring results
