+++ src/routes/index.ts
@@ -3,6 +3,7 @@ import { AppContext } from "../middleware/auth";
 import { authRoutes } from "./auth.routes";
 
 import { userRoutes } from "./users.routes";
+import { monitorRoutes } from "./monitor.route";
 
 const app = new Hono<AppContext>();
 
@@ -11,5 +12,8 @@ app.route("/auth", authRoutes);
 
 // user routes
 app.route("/users", userRoutes);
+
+// monitor routes
+app.route("/monitor", monitorRoutes);
 
 export default app;
